feat!: implemented an attack detection system
boards can now track fired shots and determine
if they hit water or ship, as well as store the
results so they can be displayed later.

also changed how the board stores data for the
occupied squares, registring not only the coords
but also the ship that was placed in there.

tests are now organized by sections.