feat!: game over and ship sets
implemented a function that returns true or false
if the game is over by checking if all the ships
have been sunk.

simplified the ship.js module to use a constructor
for setting the size of different kinds of ship
instead of using subclasses.

Each ship is individually exported so they can be
put together in custom sets.

The gameboard class now receives a set parameter
that should hold a custom ship set to be stored in
a variable called ships.