feat!: implemented Player and standardSet classes
Each Player instance has a name and a gameboard
associated to it.

An attack method was also introduced so players
can interact with each other's boards. Attacks
consummate only if the attacked square was not
attacked before.

standardSet class was created inside ship.js to
instanciate premade sets of ships for each
player's gameboard.