feat!: basic implementation of AI
Created a bot.js file containing a Bot class that
extends Player and holds methods for performing
attacks on player1's board, based on wether it hit
a target in the last round or not.

If last round was a success, it will attack one of
the adjecent squares from the previous hit, to
find the rest of the ship, being limited to the
gameboard bounds.

If last round was not a hit, or it's the first
round of the game, it will perform a random
attack.

Also reworked the test file so it can import live
instances from the code to easily execute
integration tests.