feat!: implemented a game loop that handles the turn back and forth between player and computer.
Divided game loop logic in 2 files: gameHelpers.js
for some helper functions and interactions with
the DOM module, and gameControl.js for controlling
the actual game loop.

Fixed the logic that verifies if the last attack
from the computer was able to sink a ship, so now
it only does the verification if the last attack
was a hit, in first place.