feat!: completed game loop and AI implementations, plus some bug fixes.
Reworked the structure of gameboards to store
information about occupied and attacked squares.
Each square is now represented as an object that
holds these details.

The AI now remembers the coordinates of the first
and last successful attacks on a ship, extracting
this information from the player's gameboard.

Implemented an aiming system for the AI to target
random adjacent squares from the first hit, or
random adjacent squares from the extremities of
the ship.

Now, when the AI hits a ship it seeks to target
another part of the ship in an ajdacent square.
If it misses, it reverts to the original
coordinates and tries another square from there.
However, if it succeeds, it locks it's targeting
system on that axis. So even if it misses again,
it will attempt to hit the ship on the same axis,
but in opposite direction.

If a ship is sunk at any point, the AI resumes its
search for another ship using the random attack
method.

Updated test file to mock the DOM module.

Fixed issues with game over verification.

Fixed the game loop to prevent the CPU's turn from
being called if the player's marking was invalid
(e.g., clicking on a repeated square).

Added names to ship objects for easier
identification.