feat!: Implemented a drag-and-drop feature for placing player ships on the board, along with a random placement method for the computer to position its ships.
Created setUpModal.js to handle interactions with
the dialog box, allowing players to set their
desired ship formation. The player has the option
to either reset their formation for a fresh start
or confirm it to initiate the battle.

Created dragAndDrop.js to enable the drag-and-drop
feature, allowing players to easily position their
ships on the board along the X or Y axis,
depending on the toggle button's status.

Populated HTML and CSS files with placeholder
information and basic styling to display the setup
dialog box.

Merged 'gameHelpers' and 'gameControl' files into
a single file to address timing issues when
loading the modules.

Added a recursive setFormationRandomly method to
the Bot class, that places ships in random
positions after checking the availability for each
chosen pair of coordinates.