feat: implemented result screen at the end of the game.
When the game is over a modal will pop up to
display the winner and give the player 2 options:
minimize the modal and take a look at the boards
or return to the main menu so he can play again.

Also, refactored logic for adding players to the
game in the game control.

Player and Bot classes receive a constructor
parameter to store the selected character.