Some bug fixes and a minor feature implementation
Refactored the call to the 'showSunkenShips'
function from within the battle module to prevent
errors when the strategy board is active.

Corrected an outdated line in 'place-ships.js'
that was still referencing 'getPlayerBoard'
instead of 'getHumanBoard'.

Resolved an issue with the capitalization of the
enemy's name when character quotes were called
from the Place Ships menu.

Added a feature to reveal the positions of the
CPU's ships that haven't been sunk if the CPU wins
the game.