Audio settings update pack
Modified the audio settings on the Main Menu to be
volume sliders so it's consistent with the other
pages.

Fixed an issue with radar lock alarm - audio loop
and blinking screen - where both wouldn't stop
running after the game was over.

Fixed the radar lock audio volume not being sync'd
with the sound effects volume setting.

Change radar lock audio to be a single beep with
an overall lower volume.

Fixed an issue where unset audio and language
settings would cause the game to break.