feat!: implemented a new AI that makes smarter choices instead of completely random attacks
This new AI is a protoytpe for the hard mode that
will be implemented to the game, and consists of a
decision making pattern that calculates 2 values.

The first one is the size of the biggest unsunken
ship on the player's board.

The second is the number of clear directions from
the square to be attacked, that is, the directions
where there are no previously attacked squares in
the way of a potencial ship of the size of the
first value.

Both values are recursively decreased if no square
meets the criterion.