feat!: improved and refactored AI
Renamed Bot class to AI.

Created ai-easy.js, ai-medium.js and ai-hard.js to
hold different skilled AI subclasses.

The hard mode AI now works from an array of
untouched gameboard squares that is constantly
updated to exclude the squares that have been
attacked or are adjacent to discovered ships.

Also refactored AI to be initialized with the
player set to its enemy, removing the need of
passing the player object as an argument to every
AI method.