refactor: Reworked the hard difficulty AI
Adjusted the method for assessing whether to
attack a specific square. Now, it calculates the
number of available consecutive squares in all
four directions from a random position.

By subtracting the respective values, the AI
determines if one or two axes can accommodate the
ideal target, which is the largest unsunken enemy
ship.

Refactored the AI parent class for clarity.