feat!: implemented logic for starting a new game or loading an existing one
Created navigation.js to centralize all navigation
logic, including the loading screen.

Added a confirmation step to actions that discard
the saved game data -- still using prompts, soon
to be refined.

At this stage of development, the saved game can
only be retrieved if the user refreshes the battle
page, but there will be a 'continue game' button
in the upcoming new main menu screen.