fix: fixed more issues with code being executed before user interaction
Implemented a feature in game-control and audio
modules that pauses the execution of those
segments until the user has dismissed the
"press any key" screen.

Enabled loaded games to trigger the radar lock
alarm if needed.

Fixed gameboard borders not displaying.

Implemented animation spin to the strategy board
border when all ships are placed and ready to
battle.