fix: responsive design patches
Fixed an issue where removing an ally from the
radar caused the game to break due to wrong query
selector.

Changed the screen height to use vh unit instead
of svh, to allow fullscreen on mobiles when the
browser nav bar is hidden.

Fixed fonts not being loaded on mobile devices
past the main menu page.