fix: fixed activation screen error
Modified the "press any key" screen to be
dismissed on a keypress event instead of a keydown
event. This change ensures that keys like CTRL,
SHIFT, and ALT are ignored, preventing an issue
where the browser did not recognize them as user
interactions, which caused an audio error.

Additionally, corrected a spelling error in the
CSS class name, changing "blurred" to "blurry".